Enable Idiom Recognition outside of jitserver
Idiom Recognition's set up code in TR_CISCGraph::makePreparedCISCGraphs
ensures, if a build has jitserver code enabled and the compilation
is an out-of-process compilation, that memory for the graphs it
creates is allocated using the GlobalAllocationRegion.  However, the way
the conditions are checked prevents the graphs from being created at
all if jitserver code is enabled and the compilation is an in-process
compilation.

Fixed this by ensuring the graphs are always created, but that the
GlobalAllocationRegion is in scope if the graphs are being created for
an out-of-process compilation.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>