AArch64: Implement Locals Compaction
This commit implements locals compaction on aarch64.
`mapStack` function is changed to call `mapCompactedStack` function
implemented in OMRLinkage class when locals compaction is enabled.
`mapSingleAutomatic` function is modified so that 8-byte (or larger)
alignment is applied only to stack allocated object without GC index,
because GC indices of stack allocated objects are already adjusted
in `createStackAtlas()` so that they are properly aligned.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>