Address potential race condition at checkpoint during shutdown
At shutdown, set the checkpoint status to interrupted regardless
of whether currently there is a checkpoint in progess. If a
checkpoint is in progress then the thread waiting on the CR
monitor will abort and return from the hook; otherwise, the
thread could (or is en route to) be blocked on the Comp Monitor
and so will abort before incorrectly attempting to suspend
stopping or stopped compilation threads.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>