Fix >4k field offset path for unresolved field load/store in Z codegen
Under some IL configurations the Z codegen will set the index register
for a load/store memref leaving the base register NULL. This results
in a load/store instruction using R0 as the base register. Normally
this would be a small performance problem, not a functional issue.
But for the unresolved large offset case, patching will update the base
register of the load/store  with R14 which is set to the field offset
plus the contents of R0. R0 is used because the memref base register is
NULL leaving us with NoReg, but NoReg and R0 have the same 0 definition.
In such a scenario R0 contains an undefined value and can result in
storing/loading to/form a random location.

This fix will only emit the `agr` instruction adding the base register
contents to r14 when the base register is not equal to NoReg. This
solution results in a proper load/store offset under all scenarios
including shift-by-one and shift-by>1 compressedref configurations.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>