JITServer custom metrics exporter for Prometheus
This commit implements a simple HTTP server, called MetricsServer, that
responds to "GET /metrics" requests from an agent like Prometheus.
The response includes values for all the metrics that JITServer
is willing to provide information on. This commit implements support
for four metrics:
1. CPU utilization of JITServer
2. Physical memory available to JITServer
3. Number of clients connected to JITServer
4. Number of active compilation threads in JITServer

Example of a recognized HTTP request:
```
   GET /metrics HTTP/1.1
   Host: 127.0.0.1:9403
   User-Agent: Prometheus/2.31.1
   Accept: application/openmetrics-text; version=0.0.1,text/plain;version=0.0.4;q=0.5,*/*;q=0.1
   Accept-Encoding: gzip
   X-Prometheus-Scrape-Timeout-Seconds: 3
 ```
Example of a response sent back:
```
   # HELP cpu_utilization Cpu utilization of the JITServer
   # TYPE cpu_utilization gauge
   cpu_utilization 12.000000
   # HELP available_memory Available memory for JITServer
   # TYPE available_memory gauge
   available_memory 32036204544.000000
   # HELP connected_clients Number of connected clients
   # TYPE connected_clients gauge
   connected_clients 1.000000
   # HELP active_threads Number of active compilation threads
   # TYPE active_threads gauge
   active_threads 1.000000
```
The response is in OpenMetrics format (https://openmetrics.io/) so that
other monitoring agents, beyond Prometheus, can collect metrics.

The MetricsServer implemented in this commit runs on a single, dedicated
thread. This limits the interference that the MetricsServer can have on the
rest of JITServer. Up to 4 concurrent HTTP requests can be processed
concurrently, through the use of the "poll" mechanism.
At the moment, the communication with the MetricsServer is not encrypted.
Encryption support will arrive in a future commit.

The MetricsServer functionality is disabled by default.
Command line options:
`-XX:+JITServerMetrics`            enables JITServer MetricsServer
`-XX:-JITServerMetrics`            disables JITServer MetricsServer
`-XX:JITServerMetricsPort=<NNN>`   specifies port number for MetricsServer (default 38500)

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>