Prevent pushing null appendix object for invokedynamic and invokehandle
This change impacts how the stack is set up for resolved invokedynamic
and invokehandle calls. An invokedynamic/invokehandle call is considered
resolved when the call site table entry (invokedynamic) or MethodType
table entry (invokehandle) corresponding to the bytecodes is resolved,
allowing us to check if the entries - which we refer to as invoke
cache array, contains a null appendix object. Normally, the appendix
object is implicitly pushed as the last argument for resolved
invokedynamic/invokehandle calls where we create a call for the
target method as we can determine that at compile time. However, in
some cases, the appendix object may be null, which happens when the
target method does not expect it. Therefore, we need to check if
that's the case and skip pushing the appendix object.

Furthermore, the number of args that needs to be stashed for OSR
would be determined by whether the appendix object was null.

This does not affect unresolved invokedynamic/invokehandle, as those
calls are handled via calling a VM INL method which checks if the
appendix object is null before constructing the call frame for
calling the target method.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>