Suppress inlining of recognized methods without a call node
Because of the consistency check above, we can get the recognized method
from initialCalleeMethod instead of from callNode. The only remaining
uses of the callNode and callNodeTreeTop are within the cases handling
add, subtract, and multiply for BigInteger and BigDecimal, along with
BigDecimal.valueOf. All of the other listed methods can be suppressed
without a call node. In particular this prevents TR_EstimateCodeSize
from looking for and later estimating targets for call sites that belong
to methods other than the root method, i.e. call sites within methods
that have not yet been inlined.