Add generic compilation error code for relocation failure
The generic compilation error code will be used in places that need to
determine what to do next based on the compilation failure. This is
common across most relocation failures. Therefore, it is best to have a
generic error code for this purpose, and use the relocation error code
for RAS purposes.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>