Pass TR_MethodToBeCompiled to processExceptionCommonTasks
An AOT load does not use the SharedCacheVM front end. Therefore, the
compilation failure line in the vlog did not print out whether the
compilation was a failed AOT load. This commit fixes this by passing in
the TR_MethodToBeCompiled entry, which can be used to determine whether
the compile was an AOT load or not.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>