Jenkins: Detect full url in Pull Request depends
When running a PR build with a dependent change,
if the user pastes the full URL for the other change,
the parsing will fail. This adds logic to parse a full
url in addition to existing logic that can only parse
the shorthand. Also fail if we don't recognize the
depends pattern.

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>