Always call classMatchesCachedVersion in rememberClass
See https://github.com/eclipse-openj9/openj9/issues/15013 for more
details.

It is possible for two J9Classes to have the same J9ROMClass but have
different class chains. However, because a class chain is stored into
the SCC using a key derived from the J9ROMClass, it isn't possible to
store different class chains that have the same first J9ROMClass in the
chain. This can lead to undefined behaviour as invalid AOT code can be
executed.

This commit fixes the issue by changing the code that assumes that if it
can find an existing class chain in the SCC for a given J9Class, that it
is valid for said J9Class. The code now always validates the class
chain.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>