Change to indicate how threads should suspend for checkpoint
Use SUSPEND_THREADS_FOR_CHECKPOINT to indicate that threads should be
suspended for checkpoint. The reason for this is to facilitate the
ability to compile methods proactively prior to allowing the VM to
invoke CRIU. Preparing to checkpoint then means first performing the
necesary compilation and then suspending threads. Thus, we need to be
able to distinguish between the two phases.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>