Enable DDR for platforms with no JIT compiler
DDR requires the constant INTERNAL_PTR_REG_MASK defined in
MethodMetaData.h, which was previously only included in JIT
code. This change exposes that constant to the DDR macro
extraction process so it can be included in the generated
code for DDR, even for platforms with no JIT compiler yet
(e.g. RISC-V).

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>