Notify comp threads waiting on comp monitor when preparing to checkpoint
When the compilation queue is empty, active compilation threads will
wait on the comp monitor for work. The thread that hooks into the jit to
prepare for checkpoint signals that comp threads should be suspended but
it does not notify threads waiting on the comp monitor. This can lead to
a hang where the hook thread waits to be notified by the "active" comp
thread that it is en route to suspend itself, while the "active" comp
thread waits to be notified for when there is work to do be done.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>