Release VMAccess before preparing to checkpoint
The thread that hooks into the JIT to start preparing for
checkpoint has VMAccess. However, any compilation threads currently
compiling will request Exclusive VMAccess at some point during the
compilation to add the code cache to the artifact manager. Because the
hook thread waits for all compilations in flight to complete, a deadlock
condition arises as the compilation thread waits for Exclusive VMAccess
that it can't acquire because the hook thread has VMAccess, while the
hook thread waits for the compilation thread to notify it that it will
suspend itself.

This commit fixes this by releasing VMAccess before preparing to
checkpoint, and requires it before returning to the VM. This is ok
because all other java threads have been halted at this point.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>