Skip halt check after explict GC, with external exclusive VM access
If caller of explicit GC (for example RAS) already obtained exclusive VM
access, checking if we the thread needs to halt after completing the GC
could lead to a deadlock. For example exclusive VM access could prevent
an inspector thread to proceed, while also this thread just blocked
expecting to be inspected.

The fix is to make an excpetion and skip the halt check if we know that
we hold exclusive VM access.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>