Be more confident about single interface implementer profiled guards
These guards used to be nop guards. They no longer are, but only because
certain kinds of bytecode are technically allowed to be loaded. Bytecode
that is problematic in this way should be exceedingly rare though, so
treat these guards more like the nop guards we used to have by marking
the taken side cold and by making sure to allow privatized argument
rematerialization.