Return early in String.indexOf if substring is longer than string
If substring is longer than string, then it should return -1 early since
it is impossible for the substring to be found in the original string.
This early return also avoids an assertion failure when compact strings
are enabled in the StringUTF16 indexOf helpers which checks that the
source string is longer than the substring.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>