Add class for OSR redefinition when removing a merged HCR guard
Previously, if an HCR guard was merged into another virtual guard,
it could be removed without taking into account the class whose
redefinition it guards against. If no other removed HCR guard was
sensitive to redefinition of the same class, then no redefinition
assumption would be made for that class. If that class were later
redefined, the JIT would fail to patch OSR guards within the method.

Additionally, use performTransformation() to make the removal optional
like it is in the non-merged case above.