Stop registering class extend assumptions for HCR guards
Extending the defining class of an inlined method does not invalidate
the bodies of methods defined by that class, which remain unchanged
until/unless the class is actually redefined.

It's possible for a guard merged with an HCR guard to require a class
extend assumption for its (non-HCR) virtual guard semantics. In that
case, execution will continue on and create the class extend assumption
in exactly the same way as it would without the merged HCR guard.