Correct reading BuildVersionCommand from OSX core files
It was not reading the 'platform' field causing misinterpretation
of the following fields.

The 'sdk' field was read as the number of tools yielding unreasonable
values (e.g. 11.3.0 is expressed as 0x000b0300 = 721664), with
significant expense of both time and space.

Also, the decomposition of versions was incorrect due to off-by-one
errors in the loop bounds (e.g. 11.3.0 was decomposed as 0.110.30).

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>