Enable prehashing of j.l.Thread and Class
Pre-tenured objects like Thread and Class do not get moved as often by
the GC which means that it would take longer for it to grow a hash slot.
This would require the VM to take the slow path for retrieving the
hashcode until the hash slot is grown. Pre-hashing these objects
addresses this problem.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>