Re-order the shutdown hook sequence
The new implementation of Thread.getAllThreads() makes use of method
references. This requires that a new MethodHandleNatives$CallSiteContext
is created. This registers a PhantomCleanable with the commoncleaner and
circumvents what we are attempting to achieve with
CleanerShutdown.shutdownCleaner(). We can resolve this issue re-ordering
operations such that shutdownCleaner is the last operation before we
fallback to the sending the interrupt.

Fixes: https://github.com/eclipse-openj9/openj9/issues/15453

Related: https://github.com/eclipse-openj9/openj9/issues/15408

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>