Stop trusting parameter types for bastore in DLT
The bytecode may have stored a byte[] reference into a boolean[]
parameter slot, or vice versa, on a path that does not get translated
into IL for the DLT compilation. In that case, the compiler could
incorrectly treat certain bstorei nodes as though the array type is
known based on the signature of a parameter.