Add VirtualThread native stubs and the active VirtualThread list
notifyJvmtiMountBegin() will be used to add a virtual thread to the list
and allocate thread local storage, while notifyJvmtiUnmountEnd() will be
used to remove a virtual thread from the list and deallocate thread
local storage. The list's root is a global jni ref to an empty virtual
thread to prevent references to it from becoming stale.

Issue: #15183
Signed-off-by: Eric Yang <eric.yang@ibm.com>