Update jvmtiGetThreadGroupChildren for java 19
The ThreadGroup class does not store its threads anymore and child groups
are separated into "groups" and "weaks" ThreadGroup arrays.

The child ThreadGroups array is populated from the ThreadGroup's "groups"
and "weaks" arrays, which are the child ThreadGroups that are strongly
and weakly reachable from the parent.

The threads array is constructed by iterating over all J9VMThreads and
adding the threads that belong in the given ThreadGroup.

The returned arrays are constructed while locking on the ThreadGroup
object.

Fixes: #15244
Issue: #15183
Signed-off-by: Eric Yang <eric.yang@ibm.com>