Separate codegen max precision time queries
The getSupportsCurrentTimeMaxPrecision() query is currently used by
Z to indicate to the Simplifier that it supports transforming
System.nanoTime() calls into calls to currentTimeMaxPrecision().

The getSupportsMaxPrecisionMilliTime() query is used similarly for
transforming System.currentTimeMillis(), but only if
getSupportsCurrentTimeMaxPrecision() returns true as well.

This means that we can't selectively disable the nanoTime()
transformation without disabling the currentTimeMillis() one, but we can
do the reverse.

This patch changes Simplifier to check
getSupportsCurrentTimeMaxPrecision() for nanoTime() and
getSupportsMaxPrecisionMilliTime() for currentTimeMillis() to allow
each transformation to be enabled independently of the other.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>