Add JIT System.nanoTime() CRIU test case
System.nanoTime() needs special handling in checkpoint mode. Since the
JIT has an optimized version on some platforms it needs testing to
verify that it is behaving as expected. This test case checks the JIT
against the non-JIT implementation in the simple scenario where the
reference clock moves forward between checkpoint and restore. It does
not handle the reverse case, since there's no easy way to set that up.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>