Add option to decompress() for variable length
In StringBuilder and StringBuffer, decompress() usually only works up
to the current data length. However, methods like move() can place
characters past the current length, so add a parameter to decompress()
in order to specify the number of characters that are decompressed.

Fixes issues with insert() method when compact strings is enabled and
the caller has compressed storage and the inserted object is not
compressed.

Also corrects some spelling of variable names and fixes some comments
mixing up StringBuilder and StringBuffer.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>