Try to force JIT nanoTime test method compilation
Apply count=1 to nanoTimeJit() to try to force its compilation as
early as possible so that the compiled version will be invoked
during the test, otherwise the test will be ineffective.

This solution does not guarantee compilation, but it's the biggest
hammer we have aside from `disableAsyncCompilation`, which is global
and would affect all compiled code, test run time, and all test
variations (since this option is being applied to all variations).

Note that count=1 is better than count=0 here since 0 will most likely
result in calls to System.nanoTime() being unresolved at compile time,
which prevents the optimization that we are trying to verify.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>