Remove AVX512F from the features supported by our portable CPU description
AOT code embedded in containers needs to be able to run on a wide array
of processors. For this reason, the portable AOT code is allowed to
use only a subset of processor features, subset defined by the
SandyBridge architecture. Since AVX512 instructions are not supported
by SandyBridge, the AVX512F feature should be removed from the portable
set of features.

Issue: #15697

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>