Check if instanceDescription is an immediate value in loop
There is a case where the number of object slots are greater
than the number of bits in a word and `instanceDescription`
is used as an immediate value instead of a pointer, such as
when the class doesn't have any references. We need to break
the loop in this case before `instanceDescription` is
dereferenced as a pointer.

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>