Deleted an unnecessary assert
Assert from Issue:#15481
This assert should be eliminated because it is unnecessary.
When a compilation fails with an error code, a redefinition may happen right away, marking the compilation entry as unloaded, even though the compilation has failed.
So it's totally possible to have the entry marked as unloaded and the error code be something else than compilationInterrupted.
It's quite normal to get into shouldAbortCompilation with a different error code, thus this assert is invalid and can be deleted.

Signed-off-by: Manasha Vetrivelu manasha.vetrivelu@ibm.com