Continuation Scan Update for generational GC or concurrent GC
For generational GC, local GC would not scan tenured Object, which dose
not contain any nursery reference, so if the continuation Object has
been tenured, next local GC might miss to scan it.

For GenCon and balanced GC, we need general post writebarrier to handle
the case. the barrier will be triggered after yieldContinuation.
this change does not cover concurrent Scavenger and metronome, which is
required read barrier.

Signed-off-by: Lin Hu <linhu@ca.ibm.com>