Fix Checkcast evaluator for dynamicCastClass
When a cast class is dynamic, there was a bug in checkCast evaluator
where a call to helper through runtime check in case cast class is
interface or arrayClass. Function to emit super class test in checkCast
did not account for this update, hence later we failed with a branch
instruction with NULL label symbol. This commit fixes that.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>