Update jvmtiGetStackTrace to support virtual thread
Refactors vm->internalVMFunctions to vmFuncs
Adds a continuation object parameter to jvmtiInternalGetStackTrace.
If targetThread is not null, the targetThread will be used, otherwise
check if continuation object is not null and use that.

In jvmtiGetStackTrace, adds a check for virtual thread, call
jvmtiInternalGetStackTrace with the corresponding arguments depending
on whether the thread is virtual or not.

Adds an assertion to getVMThread to ensure that targetThread will be
NULL for a virtual thread so the extra call to IS_VIRTUAL_THREAD can be
avoided.

Updates jvmtiGetFrameCount and jvmtiGetFrameLocation to remove the
unnecessary IS_VIRTUAL_THREAD call.

Updates the coding style of surrounding code

Issues: https://github.com/eclipse-openj9/openj9/issues/15759 https://github.com/eclipse-openj9/openj9/issues/15183

Signed-off-by: Gengchen Tuo <gengchen.tuo@ibm.com>