Add ability to sleep before checkpoint
This commit adds the ability to sleep before a checkpoint to allow the
compilation threads to compile the methods already present in the
compilation queue. It is important that the hook thread releases both
VM Access and the Compilation Monitor before sleeping; otherwise, the
compilation threads will not be able to empty the queue.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>