Add ability to merge multiple -Xjit: and -Xaot: options.
Specifically,multiple -Xjit: options will get merged into a single
-Xjit: option, and multiple -Xaot: options will get merged into a
single -Xaot: option. This behaviour is guarded by the options
-XX:+MergeCompilerOptions and -XX:-MergeCompilerOptions; the former
enables the behaviour and the latter disables it. By default, this
behaviour is disabled.

By default (i.e., when option merging is not enabled), if the last the
-Xjit: option (with the colon) is empty, the JVM terminates with an
error. When option merging is enabled, as long as there is a single
non-empty -Xjit: option, no error is returned. This applies for the
-Xaot: case as well.

This commit also refactors the code to deduplicate common initialization
of the -Xjit: and -Xaot: options.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>