Add j9 special args to parameters passed to addExtDir
addExtDir is used to initialize the extension directory paths,
but this function can't easily be customized based on command
line options because it has no access to that configuration
data. This commit allows addExtDir to be customized by passing
the J9SpecialArguments structure that can be tailored to hold
command-line argument information. This commit does not do
any customization; it is facilitating future customization
(or customization by downstream projects without modifying
the API for addExtDir themselves).

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>