Reinitialize target process info on restore in nonportable mode
The TR_X86ProcessorInfo struct is used in a bunch of asserts as sanity
tests to ensure consistency with the Port Library. In a restore run, in
non portable mode, the processor description acquired from the port
library can change; therefore, the TR_X86ProcessorInfo struct also needs
to be reinitialized.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>