Check whether array index variable is an induction variable
In examining an array index for Idiom Recognition, analyzeOneArrayIndex
checks whether the variable operand in a 'var + const' or 'const + var'
expression is an induction variable.  On the other hand, if the
expression is a simple variable, the analysis assumes the variable
reference is acceptable.  However, it can happen that the variable used
as the array index is not an induction variable which can result in an
incorrect transformation.

Fixed this by adding a check that an arrayindex that is a variable is
also an induction variable.

Fixes #15474

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>