Handle virtual threads in JVMTI GetOSThreadID & InterruptThread
GetOSThreadID: Throw JVMTI_ERROR_UNSUPPORTED_OPERATION for virtual
threads since this function is OpenJ9 specific, and we can dictate its
implementation.

InterruptThread: If virtual thread is unmounted i.e. targetThread is
NULL, then skip the targetThread->osThread related operations.

Due to the new synchronization for addressing the timing gap, the
virtual thread is pinned regardless whether it is mounted or not.

Fixes: #15763

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>