Fix stackwalk related jvmti functions
In the old implementation, if a carrier thread has a virtual thread
mounted after getVMThread being called, the virtual thread will be
walked. We fix this by adding a check for currentContinuation of the
carrier thread. If the currentContinuation is not null, the info of the
carrier thread has been swapped to the currentContinuation, so the
currentContinuation will be walked instead.

To facilitate other stackwalk-related functions, a generic stackwalk
helper is added to walk a platform thread or a virtual thread. A
virtual can be mounted or unmounted, if it's mounted, the carrier thread
is walked, otherwise the continuation object is walked. For a platform
thread, if the currentContinuation is not null, the currentContinuation
will be walked, otherwise the thread is walked.

Signed-off-by: Gengchen Tuo <gengchen.tuo@ibm.com>