Inspect BCI as well while computing effective caller index
In case we are in estimate code size phase of inlining or inlining was
different in previous profiling compilation, we prepare a call stack to
find out the matching call stack in the profiling information.
Previously it was just matching the method called at the call site to
find out the matching call stack. This would be incorrect as we could
have same call stack at multiple call sites and they would have
different hotness. Current solution of computing effective caller index
would simply find the first call stack that matches and does not inspect
later ones. To fix that, we should also check the byte code index at the
call site as that would provide us the correct information.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>