Revert "Preserve ymm/zmm registers on x"
This reverts PR #14632. A 10%+ performance
regression (issue #15716) reported on JDK 8
was isolated to #14632. Reverting this PR
should not have any function change, it
was only intended to support the Vector
API JEPs which have not yet been enabled
on x86 in OpenJ9. A previous attempt at
disabling just the extended vector register
preservation did not recoup the performance,
but only completely reverting this PR
restored the regression. This suggests
there is something with the way the helpers
have been restructured that is causing the
performance degradation. This PR will be
reverted while the cause is investigated.