Fix EA so that it does not assume monExit will always execute an lwsync
With Lock Reservation and Nested Locks it's possible for a monExit to
skip executing a lwsync(ppc). Yet EA disables AllocationFence nodes when
it encounters a monExit or a synchronized call before the object which
the AllocationFence is guarding escapes. This allows for cases where no
lwsync is executed before references to the object is stored to the heap
which might result in a crash when other threads attempt to read the
object contents on weak memory coherency CPUs (PPC & AArch64).

Optimizer Changes:
This fix changes EA so that it will insert new AllocationFence nodes
before monExit or volatile-access nodes that EA has decided can replace
the original AllocationFence. The code generator can then optimize the
generation of lwsync instructions so that only one lwsync is executed
and also insure that the one lwsync is unconditionally executed.

EA now keeps track of one TreeTop for each block where a monExit or
volatile-access node exists. It prefers trees with volatile-access
nodes over monExit nodes because volatile-access nodes always
unconditionally execute sync instructions where monExit nodes can skip
sync instruction in some cases.

It also removes the code that attempts to disable AllocationFence nodes
when a synchronized call proceeds the escape of an object. There is no
way to be sure that the unlock at the end of a synchronized call will
actually execute an lwsync (due to lock reservation or nested locks).
Therefore this code is unsafe and will be removed.

This change also fixes an issue where local flush elimination was not
limiting it's actions to flush nodes that were in the same block as
the allocation under consideration.

Lastly, the fix insures that AllocationFence nodes that have been
disassociated with a single specific allocation (and is therefore
serving as the fence for more then one allocation) is not considered as
a candidate for removal later on in the same EA pass. This was possible
when an earlier action in the EA pass had disassociated the fence before
an attempt to find a replacement sync node for the AllocationFence was
even attempted.

Code Generator Changes:
The PPC code generator was modified so that it would skip generating an
lwsync for AllocationFence nodes when the next tree contains a monExit
or a volatile-access node. In addition, when generating a monExit where
the previous tree contains an AllocationFence, the lwsync will be
generated in a location where it's unconditionally executed. These
changes provide a generic "peephole" like optimization that the EA
changes described above takes advantage of when it places
AllocationFence nodes.

The AArch64 code generator will need to implement similar changes in
order to optimize the number of sync instruction that are executed.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>